Módulo productos
Se validan los datos el formulario de creación
Se completa la lógica para guardar los datos en la base de datos
Se hace la lógica para que al momento de crear un producto también se cree el inventario inicial

+++ database/migrations/2022_12_19_232910_productos.php
@@ -17,7 +17,7 @@ class Productos extends Migration
             $table->increments('id_productos');
             $table->string('nombre', 50);
             $table->string('codigo', 15)->unique();
-            $table->integer('total');
+            $table->double('total');
             $table->unsignedInteger('id_unidad');
             $table->foreign('id_unidad')->references('id_unidades')->on('unidades')->onUpdate('cascade')->onDelete('cascade');
             $table->unsignedInteger('id_proveedor');
